Creation react project à la racine de app/
