+++ app/src/App.tsx
@@ -1,0 +1,11 @@
+import './App.css';
+
+function App() {
+  return (
+    <h1 className="text-3xl font-bold underline text-red-600">
+      Simple React Typescript Tailwind Sample
+    </h1>
+  );
+}
+
+export default App;
