--- app/tickets-swap/.prettierrc.json
@@ -1,4 +1,0 @@
-{
-    "tabWidth": 4,
-    "printWidth": 160
-}
