--- app/tickets-swap/config-overrides.js
@@ -1,25 +1,0 @@
-const webpack = require('webpack');
-
-module.exports = function override(config) {
-    const fallback = config.resolve.fallback || {};
-    Object.assign(fallback, {
-        crypto: require.resolve("crypto-browserify"),
-        stream: require.resolve("stream-browserify"),
-        http: require.resolve("stream-http"),
-        https: require.resolve("https-browserify"),
-        url: require.resolve("url"),
-        zlib: require.resolve("browserify-zlib"),
-        vm: require.resolve("vm-browserify")
-    })
-    config.resolve.fallback = fallback;
-    config.plugins = (config.plugins || []).concat([
-        new webpack.ProvidePlugin({
-            process: 'process/browser',
-            Buffer: ['buffer', 'Buffer']
-        })
-    ]);
-    config.devtool = false;
-    config.ignoreWarnings = [/Failed to parse source map/]
-
-    return config;
-}
