--- app/tickets-swap/package.json
@@ -1,68 +1,0 @@
-{
-  "name": "tickets-swap",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@coral-xyz/anchor": "0.30.1",
-    "@solana/spl-token": "^0.4.6",
-    "@solana/wallet-adapter-react": "0.15.35",
-    "@solana/wallet-adapter-react-ui": "0.9.35",
-    "@solana/wallet-adapter-wallets": "0.19.32",
-    "@solana/web3.js": "^1.94.0",
-    "@testing-library/jest-dom": "^5.17.0",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.5.2",
-    "@types/node": "^16.18.101",
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
-    "browserify-zlib": "^0.2.0",
-    "crypto-browserify": "^3.12.0",
-    "https-browserify": "^1.0.0",
-    "prettier": "3.3.2",
-    "react": "^18.3.1",
-    "react-app-rewired": "^2.2.1",
-    "react-dom": "^18.3.1",
-    "react-router-dom": "^6.24.1",
-    "react-scripts": "5.0.1",
-    "stream-browserify": "^3.0.0",
-    "stream-http": "^3.2.0",
-    "typescript": "^4.9.5",
-    "url": "^0.11.3",
-    "vm-browserify": "^1.1.2",
-    "web-vitals": "^2.1.4"
-  },
-  "scripts": {
-    "start": "react-app-rewired start",
-    "build": "react-app-rewired build",
-    "test": "react-app-rewired test",
-    "eject": "react-scripts eject",
-    "lint": "npx eslint './src/**/*.{js,jsx,ts,tsx}'",
-    "lint-fix": "npx eslint './src/**/*.{js,jsx,ts,tsx}' --fix",
-    "fix": "npx prettier --write './src/**/*.{js,jsx,ts,tsx,css,scss}'",
-    "fix-all": "npm run fix && npm run lint-fix"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "devDependencies": {
-    "autoprefixer": "^10.4.19",
-    "postcss": "^8.4.39",
-    "tailwindcss": "^3.4.4"
-  }
-}
