--- app/tickets-swap/postcss.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-    plugins: {
-      tailwindcss: {},
-      autoprefixer: {},
-    },
-  }
-
