--- app/tickets-swap/public/manifest.json
@@ -1,15 +1,0 @@
-{
-  "short_name": "Tickets Swap",
-  "name": "Tickets Swap",
-  "icons": [
-    {
-      "src": "favicon.ico",
-      "sizes": "16x16",
-      "type": "image/x-icon"
-    }
-  ],
-  "start_url": ".",
-  "display": "standalone",
-  "theme_color": "#000000",
-  "background_color": "#ffffff"
-}
