--- app/tickets-swap/src/index.tsx
@@ -1,33 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-import "./index.css";
-import App from "./App";
-import reportWebVitals from "./reportWebVitals";
-import { ConnectionProvider, WalletProvider } from "@solana/wallet-adapter-react";
-import { WalletModalProvider } from "@solana/wallet-adapter-react-ui";
-import { PhantomWalletAdapter } from "@solana/wallet-adapter-wallets";
-import { clusterApiUrl } from "@solana/web3.js";
-import "@solana/wallet-adapter-react-ui/styles.css"; // Import default styles
-
-//const network = clusterApiUrl("devnet");
-const network = "http://127.0.0.1:8899"; // Utilisez l'URL de votre validateur local
-
-const wallets = [new PhantomWalletAdapter()];
-
-const root = ReactDOM.createRoot(document.getElementById("root") as HTMLElement);
-root.render(
-    <React.StrictMode>
-        <ConnectionProvider endpoint={network}>
-            <WalletProvider wallets={wallets} autoConnect>
-                <WalletModalProvider>
-                    <App />
-                </WalletModalProvider>
-            </WalletProvider>
-        </ConnectionProvider>
-    </React.StrictMode>,
-);
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
