--- app/tickets-swap/src/pages/CreateEvent.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-function CreateEvent() {
-    return <div>Page CreateEvent...</div>;
-}
-
-export default CreateEvent;
