--- app/tickets-swap/src/pages/SwapTicket.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-function CreateEvent() {
-    return <div>Page SwapTicket...</div>;
-}
-
-export default CreateEvent;
